orangepi_bridge/requirements.txt
# MQTT 客戶端
paho-mqtt==1.6.1

# HTTP 請求庫
requests==2.31.0
